Added a few new particles. Last point in four momentum class where I have doubles and not long doubles

+++ FourMomentum.h
@@ -8,12 +8,16 @@
 class FourMomentum 
 {
 private:
-  double energy;     // Energy component E
-  double px, py, pz; // Momentum components along the x, y, and z axes
+  // double energy;     // Energy component E
+  long double energy;     // Energy component E
+
+  // double px, py, pz; // Momentum components along the x, y, and z axes
+  long double px, py, pz; // Momentum components along the x, y, and z axes
 
 public:
   // Constructor to initialize the four-momentum components
-  FourMomentum(double energy = 0.0, double px = 0.0, double py = 0.0, double pz = 0.0, bool energy_is_rest_mass = false);
+  // FourMomentum(double energy = 0.0, double px = 0.0, double py = 0.0, double pz = 0.0, bool energy_is_rest_mass = false);
+  FourMomentum(long double energy = 0.0, long double px = 0.0, long double py = 0.0, long double pz = 0.0, bool energy_is_rest_mass = false);
 
   // Copy constructor, Move constructor, Copy assignment operator, Move assignment operator, and Destructor
   FourMomentum(const FourMomentum &other);                // Copy constructor
@@ -40,7 +44,8 @@ public:
   // Function to calculate the invariant mass (magnitude) of the four-momentum
   double invariant_mass() const;
   // Function to perform Lorentz boost to four-momentum
-  void lorentz_boost(double v_x, double v_y, double v_z);
+  // void lorentz_boost(double v_x, double v_y, double v_z);
+  void lorentz_boost(long double v_x, long double v_y, long double v_z);
   void lorentz_boost(std::vector<double> v_xyz);
 
   // Overloaded operators
